# swagger.json
{
  "info": {
    "title": "My Nifty Insight 2",
    "version": "1.0.0",
    "description": "This describes an Insight Module that can be added to Murano Exchange.  An Insight Module can implement one or more Insight Functions.  These Insight Functions can then be added to Assets as a Transformation.",
    "contact": {
      "name": "Your Name Here",
      "email": "YOU@BIZ.com"
    }
  },
  "host": "http://undefined:undefined",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "apis": [
    "./src/route*.js",
    "./swagger/definitions.yaml"
  ],
  "swagger": "2.0",
  "paths": {
    "/info": {
      "get": {
        "operationId": "listInsights",
        "description": "Get some info about this Insight",
        "responses": {
          "200": {
            "description": "Insight Functions successfully listed",
            "schema": {
              "$ref": "#/definitions/InsightInfoResults"
            }
          },
          "default": {
            "description": "Errors",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/insight/{function_id}": {
      "get": {
        "operationId": "infoInsight",
        "description": "Get info about one Insight Function",
        "responses": {
          "200": {
            "description": "Insight Functions successfully listed",
            "schema": {
              "$ref": "#/definitions/InsightInfo"
            }
          },
          "default": {
            "description": "Errors",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/insights": {
      "post": {
        "operationId": "listInsights",
        "description": "Get a list of availible Insight Functions and info about them",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Get a list of availible insight functions",
            "required": true,
            "schema": {
              "$ref": "#/definitions/InsightsFilterParams"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Insight Functions successfully listed",
            "schema": {
              "$ref": "#/definitions/InsightListResults"
            }
          },
          "default": {
            "description": "Errors",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/process": {
      "post": {
        "operationId": "process",
        "description": "Your function to process a bunch of Signal Data.\nIt is good form to copy all the fields of a source SignalData into the new SignalData,\nthen update the properties according to your function.\n",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Data to process and arguments on how to process it",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SignalDataObjectArray"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Data successfully",
            "schema": {
              "$ref": "#/definitions/SignalDataArrayArray"
            }
          },
          "default": {
            "description": "Error that stops the pipeline. Non-fatal errors should produce new signals on a seperate outlet.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "InsightInfoResults": {
      "type": "object",
      "description": "Info about this Insight Module",
      "properties": {
        "group_id_required": {
          "type": "boolean",
          "description": "True if the listInsights requires group_id to be something to list functions."
        },
        "name": {
          "type": "string",
          "description": "Presented name"
        },
        "description": {
          "type": "string",
          "description": "Additional text about this insight"
        },
        "wants_lifecycle_events": {
          "type": "boolean",
          "description": "True if this insight wants to be notified when linkages using it are created and deleted.\nDefaults to false.\n"
        }
      },
      "required": [
        "group_id_required"
      ]
    },
    "InsightsFilterParams": {
      "type": "object",
      "description": "Filter options for insight functions",
      "properties": {
        "group_id": {
          "type": "string",
          "description": "Which group of insight functions to get.  The meaning of the contents is up to the Insight Module.\nIncluding ignoring it and always returning all.\n"
        },
        "limit": {
          "type": "integer",
          "description": "Limit how many insights to fetch"
        },
        "offset": {
          "type": "integer",
          "description": "Offset into the returned insights"
        }
      },
      "required": [
        "group_id"
      ]
    },
    "SignalTypeInputInfo": {
      "type": "object",
      "description": "What data type the incoming Signal must have. If omitted, than allows any type.",
      "required": [
        "data_type",
        "name"
      ],
      "properties": {
        "data_type": {
          "type": "string",
          "description": "The type that the connecting signal must have"
        },
        "data_unit": {
          "type": "array",
          "description": "The units that the connecting signal must have",
          "items": {
            "type": "string",
            "description": "A required unit."
          }
        },
        "tag": {
          "type": "string",
          "description": "Optional tag to add to the SingalData for inlet matching"
        },
        "name": {
          "type": "string",
          "description": "Name of this signal"
        },
        "description": {
          "type": "string",
          "description": "Optional text to help users."
        }
      }
    },
    "SignalTypeOutputInfo": {
      "type": "object",
      "description": "What data type the incoming Signal must have. If omitted, than allows any type.",
      "properties": {
        "data_type": {
          "type": "string",
          "description": "The type that the connecting signal must have"
        },
        "data_unit": {
          "type": "string",
          "description": "The unit that the connecting signal must have"
        }
      },
      "required": [
        "data_type"
      ]
    },
    "ConstantInfo": {
      "type": "object",
      "description": "A constant that needs to be provided by the user",
      "properties": {
        "name": {
          "type": "string",
          "description": "The name of this constant"
        },
        "type": {
          "type": "string",
          "enum": [
            "string",
            "number"
          ],
          "description": "What kind of value this constant accepts"
        },
        "description": {
          "type": "string",
          "description": "Some words for the user about this constant"
        },
        "default": {
          "type": [
            "string",
            "number"
          ],
          "description": "Optional value to use if user doesn't specify one."
        },
        "enum": {
          "type": "array",
          "description": "An array of allowed values."
        },
        "maximum": {
          "type": "number",
          "description": "For number type, the largest it can be"
        },
        "minimum": {
          "type": "number",
          "description": "For number type, the smallest it can be"
        }
      },
      "required": [
        "name",
        "type"
      ]
    },
    "InsightInfo": {
      "type": "object",
      "description": "One insight function",
      "properties": {
        "id": {
          "type": "string",
          "description": "The unique id for this insight function"
        },
        "name": {
          "type": "string",
          "description": "A human friendly name for this insight function"
        },
        "description": {
          "type": "string",
          "description": "A description of this insight function to display to users."
        },
        "type": {
          "type": "string",
          "enum": [
            "transform",
            "rule",
            "action"
          ],
          "description": "What kind of insight this function is.  Determines how it gets used and how the UI\npresents it.  If omitted, then 'transform' is used.\n"
        },
        "asynchronous": {
          "type": "boolean",
          "description": "When true, this function requires the Callback Info block since it will respond\nwith some or all of its results asynchronously.\n"
        },
        "history": {
          "type": "object",
          "description": "When this object is present, a timeseries query will be called and the results sent along\nto the function call.  The properties here will be used to augment the query.  Only the\nSignals connected to the calling of this function will be queried.\n",
          "additionalProperties": {
            "type": "object",
            "description": "An additional parameter to the query.  Use only one property. (value, constant, template)",
            "properties": {
              "value": {
                "type": [
                  "string",
                  "number"
                ],
                "description": "The exact value to use"
              },
              "constant": {
                "type": "string",
                "description": "Which of this function's constants to use the value of here"
              },
              "template": {
                "type": "string",
                "description": "Which of this function's constants to use the value of here, fitted into a templeate\nstring.  The constants must be surrounded with '{{}}'.  For example '-{{days}}d'\n"
              }
            }
          }
        },
        "constants": {
          "type": "array",
          "description": "Additional user specified values to pass to this insight function",
          "items": {
            "$ref": "#/definitions/ConstantInfo"
          }
        },
        "inlets": {
          "type": "array",
          "description": "The kinds of data that can be connected. (This is an unordered set)",
          "maxItems": 5,
          "items": {
            "$ref": "#/definitions/SignalTypeInputInfo"
          }
        },
        "outlets": {
          "$ref": "#/definitions/SignalTypeOutputInfo"
        }
      },
      "required": [
        "id",
        "name",
        "description"
      ]
    },
    "InsightListResults": {
      "type": "object",
      "description": "The found insight functions with count info",
      "properties": {
        "total": {
          "type": "integer",
          "description": "The total number of insight functions in this group"
        },
        "count": {
          "type": "integer",
          "description": "Number of insight functions returned"
        },
        "insights": {
          "type": "array",
          "description": "The insight functions",
          "items": {
            "$ref": "#/definitions/InsightInfo"
          }
        }
      },
      "required": [
        "insights",
        "total",
        "count"
      ]
    },
    "LifecycleEvent": {
      "type": "object",
      "description": "A Lifecycle event",
      "properties": {
        "event": {
          "type": "string",
          "description": "What action caused this event",
          "pattern": "(create|delete)"
        },
        "id": {
          "type": "string",
          "description": "The unique id for the linkage that has the event.  This id will match in the calls to process."
        },
        "args": {
          "type": "object",
          "description": "The arguments for this linkage that will be also passed to each call of process."
        }
      }
    },
    "SignalData": {
      "type": "object",
      "description": "A single instance of a piece of data inside a signal.",
      "properties": {
        "tags": {
          "type": "object",
          "description": "Tag-Value pairs to help ID the data. Used to tie the data back to an\nAsset or Device or other things.\n",
          "additionalProperties": {
            "type": "string",
            "description": "Value for a tag",
            "minLength": 1
          }
        },
        "ts": {
          "type": "integer",
          "description": "Unix timestamp in microseconds of when the data originated"
        },
        "gts": {
          "type": "integer",
          "description": "Unix timestamp in microseconds of when this SignalData was generated"
        },
        "value": {
          "type": [
            "string",
            "number",
            "object"
          ],
          "description": "The value for this instance of data"
        },
        "origin": {
          "type": "string",
          "description": "The original Publishing ID",
          "minLength": 1
        },
        "generated": {
          "type": "string",
          "description": "The Publishing ID that created this SignalData.",
          "minLength": 1
        },
        "ttl": {
          "type": "integer",
          "description": "Value used to help prevent data from infinitely cycling"
        }
      },
      "required": [
        "ts",
        "value",
        "origin",
        "generated"
      ]
    },
    "HistoricalData": {
      "type": "object",
      "description": "The history of signals.  Each signal id is the key, with an array of timestamp, value pairs.",
      "additionalProperties": {
        "type": "array",
        "description": "The history of signals.  Each signal id is the key, with an array of timestamp, value pairs.",
        "items": {
          "type": "object",
          "description": "One point in history",
          "properties": {
            "ts": {
              "type": "integer",
              "description": "Unix timestamp in microseconds of when the data was stored"
            },
            "value": {
              "type": [
                "string",
                "number",
                "object"
              ],
              "description": "The value for this instance of data"
            },
            "tags": {
              "type": "object",
              "description": "Tag-Value pairs to help ID the data.  (Tags here are a subset of those in SignalData)",
              "additionalProperties": {
                "type": "string",
                "description": "Value for a tag",
                "minLength": 1
              }
            }
          },
          "required": [
            "ts",
            "value"
          ]
        }
      }
    },
    "CallbackInfo": {
      "type": "object",
      "description": "URL and auth token to use to send results if not immediately available.",
      "properties": {
        "url": {
          "type": "string",
          "description": "The URL to post the results to."
        },
        "token": {
          "type": "string",
          "description": "The token to put in the AUTHORIZATION header."
        }
      }
    },
    "SignalDataArray": {
      "type": "array",
      "description": "Array of Signal Data",
      "items": {
        "$ref": "#/definitions/SignalData"
      }
    },
    "SignalDataArrayArray": {
      "type": "array",
      "description": "Array of Array of Signal Data.  An array for each outlet, each outlet can have many signals.",
      "items": {
        "$ref": "#/definitions/SignalDataArray"
      }
    },
    "SignalDataObjectArray": {
      "type": "object",
      "description": "The data and arguments to be processed.",
      "properties": {
        "id": {
          "type": "string",
          "description": "A unique ID for this instance of this function in an asset pipeline."
        },
        "data": {
          "$ref": "#/definitions/SignalDataArray"
        },
        "history": {
          "$ref": "#/definitions/HistoricalData"
        },
        "cbi": {
          "$ref": "#/definitions/CallbackInfo"
        },
        "args": {
          "type": "object",
          "description": "Object of instance arguments.",
          "properties": {
            "function_id": {
              "type": "string",
              "description": "Which insight function to run"
            },
            "group_id": {
              "type": "string",
              "description": "Which group_id this is under"
            },
            "constants": {
              "type": "object",
              "description": "Additional constant parameters for the selected function"
            }
          }
        }
      }
    },
    "ErrorResponse": {
      "type": "object",
      "description": "Error response that stops the pipeline.",
      "properties": {
        "message": {
          "type": "string",
          "description": "Error message"
        },
        "status": {
          "type": "number",
          "description": "A numerical code for the error"
        },
        "name": {
          "description": "Error type",
          "type": "string"
        }
      }
    }
  },
  "responses": {},
  "parameters": {},
  "securityDefinitions": {},
  "tags": []
}